feat(ui): UnitAttackPower and UnitRangedAttackPower, and fix two mistyped fields
Both return THREE values, not one, and the fields behind them were wrong in
frozen's descriptor in two ways that would have produced confident nonsense:

attackPowerMods is a PACKED PAIR of int16 in a single dword -- positive low,
negative high. The reference reads it as two shorts off 0x1d8 and 0x1da. It was
declared as one int32_t, so a caller reading it whole would have got both
modifiers mashed into one number.

attackPowerMultiplier is a FLOAT living in a block whose other fields are
integers, and it was declared int32_t. Reading it as an integer yields the raw
bit pattern -- a multiplier of about 1.07e9 for a perfectly ordinary 1.0f.
Both multipliers are corrected to float; nothing else in the tree read them, so
the change is contained.

The scale applied is the multiplier PLUS ONE (the constant at 009e1130 is
1.0f), so a unit with no multiplier reports its raw values rather than zero.
Rounding is x87 round-to-nearest in the reference, so nearbyint rather than
half-away-from-zero.

Written out twice rather than shared: the reference has two separate functions,
and a common helper would put a call in frozen's sequence that is not in the
reference's.

The gate fired again -- 1 lost by callgraph -- and again the matcher was moving
a wrong link rather than losing a right one. 0060c1f0 had CGObject::GetGUID on
it and swapped to CGUnit::Unit when these bindings changed a call sequence. It
is neither: it resolves a unit token to a guid, asks the object manager for it
with TYPEMASK_UNIT, and names ScriptEvents.cpp in its own assert string. That is
Script_GetUnitFromName, now pinned.

Recorded as open in the recomp README rather than papered over: CGUnit::Unit is
still parked on 00513c30, which is a three-way party/raid/battlefield lookup.
CGUnit::Unit is a one-line accessor the reference inlines everywhere, so it
likely has no standalone address and every link for it is spurious -- but
overrides.json can only correct a link by naming a frozen counterpart, and there
is none to name. The fix is a form that blocks an address without claiming one.

vs HEAD: 0 lost, 0 gained, one link corrected. ported 2176->2178, stubs 511->509.
Built clean; not seen running.

+++ src/object/client/CGUnit.hpp
@@ -63,11 +63,19 @@ struct CGUnitData {
     int32_t baseHealth;
     int32_t pad3;
     int32_t attackPower;
+
+    // A PACKED PAIR, not one number: the positive modifier is the low int16 and the negative one
+    // the high int16. The reference reads them as two shorts off 0x1d8 and 0x1da.
     int32_t attackPowerMods;
-    int32_t attackPowerMultiplier;
+
+    // A float, in a block whose fields are otherwise integers. It was declared int32_t here, which
+    // would have reported the raw bit pattern as a multiplier. The scale callers want is this
+    // PLUS ONE, so zero means unscaled.
+    float attackPowerMultiplier;
+
     int32_t rangedAttackPower;
     int32_t rangedAttackPowerMods;
-    int32_t rangedAttackPowerMultiplier;
+    float rangedAttackPowerMultiplier;
     float minRangedDamage;
     float maxRangedDamage;
     int32_t powerCostModifier[7];
